Update InstanceStats.java
change app memory to Integer

+++ cloudfoundry-client-lib/src/main/java/org/cloudfoundry/client/lib/domain/InstanceStats.java
@@ -49,7 +49,7 @@ public class InstanceStats {
 			return disk;
 		}
 
-		public double getMem() {
+		public Integer getMem() {
 			return mem;
 		}
